fix: add missing exports
