+++ src/network/tags.ts
@@ -6,13 +6,27 @@
 
 import { getSoapFetch } from './fetch';
 import { JSNS, SHELL_APP_ID } from '../constants';
-import type {
-	CreateTagRequest,
-	CreateTagResponse,
-	TagActionRequest,
-	TagActionResponse
-} from '../types/network';
+import type { SoapBody } from '../types/network';
 import type { Tag } from '../types/tags';
+
+export type CreateTagRequest = SoapBody<{
+	tag: Omit<Tag, 'id'>;
+}>;
+export type CreateTagResponse = {
+	tag: [Tag];
+};
+export type TagActionRequest = SoapBody<{
+	action: {
+		op: 'rename' | 'color' | 'delete' | 'update';
+		id: string;
+		name?: string;
+		color?: number;
+		rgb?: string;
+	};
+}>;
+export type TagActionResponse = SoapBody<{
+	action: { op: string; id: string };
+}>;
 
 export const createTag = (tag: Omit<Tag, 'id'>): Promise<CreateTagResponse> =>
 	getSoapFetch(SHELL_APP_ID)<CreateTagRequest, CreateTagResponse>('CreateTag', {
